Create a DbSourceDiscoverUtil for common utilities
